pyproject.toml
[build-system]
requires = ["setuptools>=65.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "multi-agent-security"
version = "0.1.0"
description = "FP-15: Multi-Agent Security Testing Framework"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "scipy>=1.10",
    "networkx>=3.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov"]
